% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.autoiml_as_list}
\alias{.autoiml_as_list}
\title{Safely convert environment to list, pass through lists and handle NULL}
\usage{
.autoiml_as_list(x)
}
\description{
Safely convert environment to list, pass through lists and handle NULL
}
\keyword{internal}
