% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.autoiml_ensure_model}
\alias{.autoiml_ensure_model}
\title{Ensure final model exists in ctx, training if needed with clear warning}
\usage{
.autoiml_ensure_model(ctx, task)
}
\description{
Ensure final model exists in ctx, training if needed with clear warning
}
\keyword{internal}
