% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.autoiml_extdata_path}
\alias{.autoiml_extdata_path}
\title{Resolve a file in inst/extdata with robust package/dev fallback.}
\usage{
.autoiml_extdata_path(file)
}
\description{
Resolve a file in inst/extdata with robust package/dev fallback.
}
\keyword{internal}
