% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.autoiml_framework_requirements}
\alias{.autoiml_framework_requirements}
\title{Load framework requirements matrix from inst/extdata.}
\usage{
.autoiml_framework_requirements()
}
\description{
Load framework requirements matrix from inst/extdata.
}
\keyword{internal}
