% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.autoiml_purposes}
\alias{.autoiml_purposes}
\title{Valid claim purposes}
\format{
An object of class \code{character} of length 4.
}
\usage{
.autoiml_purposes
}
\description{
Valid claim purposes
}
\keyword{internal}
