% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.autoiml_read_yaml_file}
\alias{.autoiml_read_yaml_file}
\title{Read YAML file if yaml package is available.}
\usage{
.autoiml_read_yaml_file(path)
}
\description{
Read YAML file if yaml package is available.
}
\keyword{internal}
