% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.autoiml_validate_framework_requirements}
\alias{.autoiml_validate_framework_requirements}
\title{Validate shape of framework requirements YAML.}
\usage{
.autoiml_validate_framework_requirements(x)
}
\description{
Validate shape of framework requirements YAML.
}
\keyword{internal}
