% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.autoiml_validate_iel_rules}
\alias{.autoiml_validate_iel_rules}
\title{Validate shape of IEL rules YAML.}
\usage{
.autoiml_validate_iel_rules(x)
}
\description{
Validate shape of IEL rules YAML.
}
\keyword{internal}
